+++ src/read.jl
@@ -1544,7 +1544,7 @@ Read tabular data from an Excel file, `source`, and return it as a `Tables.jl` c
 The resulting table object can be passed directly to any function that accepts `Tables.jl` data 
 (e.g. `DataFrame` from package `DataFrames.jl`).
 
-This function requires FileIO.jl to be active in the current environment.
+This function requires both FileIO.jl v1.20.0 or higher to be active in the current environment and a Julia version >= v1.9.
 
 #### Arguments:
 
@@ -1600,7 +1600,7 @@ end
 ```
 Save a `Tables.jl` compatible table to an Excel file, `source`.
 
-This function requires FileIO.jl to be active in the current environment.
+This function requires both FileIO.jl v1.20.0 or higher to be active in the current environment and a Julia version >= v1.9.
 
 #### Arguments: